requirements.txt
fastapi
uvicorn
starlette
requests
beautifulsoup4
dash
dash-cytoscape
python-dotenv
aiocache
pytest
pytest-mock
